by ravi

+++ visog-job-portal-ui/WebContent/resources/js/address.js
@@ -88,12 +88,14 @@
 
 						// HTTP DELETE- delete role by Id
 						$scope.deleteAddress = function(address) {
+							if(confirm("Are you sure to Delete ?") == true){
 							$http(
 									{
 										method : 'DELETE',
 										url : 'http://localhost:8080/visog-job-portal-api/transaction/address/'
 												+ address.id
 									}).then(_success, _error);
+							}
 						};
 
 						// In case of edit, populate form fields and assign form.id
