+++ visog-job-portal-ui/WebContent/resources/js/files.js
@@ -73,12 +73,14 @@
 
 						// HTTP DELETE- delete role by Id
 						$scope.deleteFiles = function(files) {
+							if(confirm("Are you sure to Delete ?") == true){
 							$http(
 									{
 										method : 'DELETE',
 										url : 'http://localhost:8080/visog-job-portal-api/transaction/files/'
 												+ files.id
 									}).then(_success, _error);
+							}
 						};
 
 						// In case of edit, populate form fields and assign form.id
